+++ HISTORY.md
@@ -7,9 +7,15 @@
 - Generalize use of working directories to support runs on S3 mounted
   filesystems. Ensures all work takes place inside transactional directories.
   Thanks to Tetiana Khotiainsteva and Sven-Eric Schelhorn.
+- Provide separate germline calling for somatic tumor/normal pairs. Supplements
+  somatic calls with standard germline calls on normal samples, including
+  ensemble and SV calling.
 - Support FreeBayes 1.1.0 for improved memory usage and 3-4x speedup.
   Will default to 1.1.0 at next release. Validation work:
   https://github.com/bcbio/bcbio.github.io/blob/master/_posts/2016-11-21-giab-hg38-freebayes.md
+- Rework quality control for speed and output directory consistency. Avoid
+  re-duplicating calculations and put all output in qc directory to make re-runs
+  easier. Thanks to Vlad Saveliev.
 - Fixes for Seq2C concurrency problems when preparing BED files. Thanks to Vlad
   Saveliev.
 - Support for VEP 86 with additional plugins and generalization of fields.